School Project 2
Online store project using MySQL database "phpstore". Allows items to be
added and deleted from a cart. The database has a few users and two
superusers. The users can submit their cart for processing and then view
their orders. The superusers, can process(delete) orders and also add
and modify items. All non-users, users, and superusers can create a
cart. For the sake of simplicity, the password for the users is their
name.